fixed some minor issues. Modified the way for sending message via java code. Added send test message from custom plugin(now only in load method)
