+++ drill-agent/src/nativeAgentMain/kotlin/com/epam/drill/core/SymbolsRegister.kt
@@ -2,7 +2,12 @@
 
 package com.epam.drill.core
 
+import com.epam.drill.api.sendToSocket
 import com.epam.drill.jvmapi.JNIEnvPointer
+import com.epam.drill.toKString
+import jvmapi.JNIEnv
+import jvmapi.jobject
+import jvmapi.jstring
 import jvmapi.jvmtiError
 
 @CName("currentEnvs")
@@ -14,4 +19,9 @@ fun currentEnvs(): JNIEnvPointer {
 @CName("checkEx")
 fun checkEx(errCode: jvmtiError, funName: String): jvmtiError {
     return com.epam.drill.jvmapi.checkEx(errCode, funName)
+}
+
+@CName("Java_com_epam_drill_plugin_api_processing_Sender_sendMessage")
+fun sendFromJava(env: JNIEnv, thiz: jobject, message: jstring) {
+    sendToSocket(message.toKString() ?: "empty...")
 }
