+++ drill-plugin-api/drill-agent-part/src/drillAgentPartMain/kotlin/com/epam/drill/plugin/api/processing/Sender.kt
@@ -1,0 +1,5 @@
+package com.epam.drill.plugin.api.processing
+
+object Sender {
+    external fun sendMessage(message: String)
+}
