--- plugins/drill-custom-plugin-native/.gitignore
@@ -1,4 +1,0 @@
-build/
-/.gradle/
-/drill-api/mingwX64/main.dll
-/drill-api/mingwX64/drill4j.klib
