--- plugins/drill-custom-plugin-native/build.gradle
@@ -1,72 +1,0 @@
-import org.apache.tools.ant.taskdefs.condition.Os
-import org.jetbrains.kotlin.gradle.plugin.mpp.NativeBuildType
-import org.jetbrains.kotlin.gradle.plugin.mpp.NativeOutputKind
-
-plugins {
-    id 'kotlin-multiplatform'
-    id 'kotlinx-serialization'
-}
-repositories {
-    mavenCentral()
-    mavenLocal()
-    jcenter()
-    maven { url "https://dl.bintray.com/kodein-framework/Kodein-DI" }
-    maven { url "https://dl.bintray.com/soywiz/soywiz" }
-    maven { url "https://dl.bintray.com/kotlin/kotlin-eap" }
-    maven { url "https://kotlin.bintray.com/ktor" }
-    maven { url "https://dl.bintray.com/spekframework/spek-dev" }
-    maven { url "https://kotlin.bintray.com/kotlinx" }
-    maven { url "https://drill4j.bintray.com/nativewebsockets" }
-}
-kotlin {
-    targets {
-        fromPreset(presets.jvm, 'adminPart')
-        fromPreset(presets.jvm, 'agentPart')
-    }
-    sourceSets {
-        agentPartMain {
-            dependencies {
-                //fixme 6 only..
-                implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8")
-                implementation project(":drill-plugin-api:drill-agent-part")
-                implementation project(":drill-common")
-            }
-        }
-        adminPartMain {
-            dependencies {
-                implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8")
-                implementation project(":drill-plugin-api:drill-admin-part")
-                implementation project(":drill-common")
-            }
-        }
-
-    }
-}
-
-
-
-adminPartJar {
-    archiveName = "admin-part.jar"
-    from(file("./plugin_config.json")) {
-        into("static")
-    }
-
-}
-agentPartJar {
-    archiveName = "agent-part.jar"
-    from(file("./plugin_config.json")) {
-        into("static")
-    }
-}
-
-task packPlugin(type: Jar) {
-    from(adminPartJar, agentPartJar)
-    destinationDir = file("../../distr/adminStorage")
-}
-
-
-task buildCustomPlugin {
-    group = "app"
-    dependsOn "packPlugin"
-
-}
