--- plugins/drill-custom-plugin-native/plugin_config.json
@@ -1,4 +1,0 @@
-{
-  "id": "custom",
-  "apiVersion": "v1"
-}
