--- plugins/drill-custom-plugin-native/src/agentPartMain/kotlin/com/epam/drill/plugins/custom/AgentPart.kt
@@ -1,21 +1,0 @@
-package com.epam.drill.plugins.custom
-
-import com.epam.drill.plugin.api.processing.AgentPluginPart
-
-
-@Suppress("unused")
-/**
- * @author Denis Domashenko on 2/22/19.
- */
-class AgentPart(override val id: String) : AgentPluginPart() {
-    override fun load() {
-        println("Plugin $id loaded")
-    }
-
-
-    override fun unload() {
-        println("JAVA SIDE: Plugin '$id' unloaded")
-    }
-
-
-}
