+++ plugins/drill-custom-plugin/plugin_config.json
@@ -1,0 +1,4 @@
+{
+  "id": "custom",
+  "apiVersion": "v1"
+}
