+++ plugins/drill-custom-plugin/settings.gradle
@@ -1,0 +1,20 @@
+pluginManagement {
+    resolutionStrategy {
+        eachPlugin {
+            if (requested.id.id == "kotlin-multiplatform") {
+                useModule("org.jetbrains.kotlin:kotlin-gradle-plugin:1.3.20-eap-100")
+            }
+            if (requested.id.id == "kotlinx-serialization") {
+                useModule("org.jetbrains.kotlin:kotlin-serialization:1.3.20-eap-100")
+            }
+        }
+    }
+    repositories {
+        mavenLocal()
+        mavenCentral()
+        maven { url 'https://plugins.gradle.org/m2/' }
+        maven { url "https://dl.bintray.com/kotlin/kotlin-eap" }
+    }
+}
+rootProject.name = 'drill-custom-plugin'
+
